Extract BaseGitRepositoryBackedTest
