+++ backendImpl/src/test/resources/setup-for-diverged-and-older-than.sh
@@ -2,13 +2,13 @@
 
 set -e -o pipefail -u
 
-source ./common.sh
+self_dir=$(cd "$(dirname "$0")" &>/dev/null; pwd -P)
+source "$self_dir"/common.sh
 
 push() {
   b=$(git symbolic-ref --short HEAD)
   git push -u origin $b
 }
-
 
 newrepo $1 machete-sandbox-remote --bare
 newrepo $1 machete-sandbox
