+++ backendImpl/src/test/resources/setup-with-single-remote.sh
@@ -10,7 +10,6 @@ push() {
   git push -u origin $b
 }
 
-
 newrepo $1 machete-sandbox-remote --bare
 newrepo $1 machete-sandbox
