+++ src/main/resources/META-INF/plugin.xml
@@ -103,7 +103,7 @@
 
         <!-- TODO (#274): consider excluding this action from the production builds of the plugin -->
         <action id="GitMachete.OpenTabAction"
-                class="com.virtuslab.gitmachete.frontend.actions.ideprobe.OpenTabAction"
+                class="com.virtuslab.gitmachete.frontend.actions.unanchored.OpenTabAction"
                 text="Git Machete: Open Tab"/>
     </actions>
 </idea-plugin>
